Boomerang tuning - egui UI
